requirements.txt

flask
pillow
numpy
tensorflow
